+++ src/board/mod.rs
@@ -42,7 +42,7 @@ pub struct Undo {
 pub struct Board {
     pub occupancies: [u64; 3],
     pub bitboards: [u64; Piece::COUNT],
-    pub king_square: [u64; 2],
+    pub king_square: [usize; 2],
     pub pieces: [Piece; 64],
     pub side: Color,
 
@@ -169,7 +169,7 @@ impl Board {
             set_bit(&mut position.occupancies[Color::BOTH.index()], square);
             position.pieces[square] = piece;
             if piece.piece_type() == PieceType::KING {
-                position.king_square[piece.color().index()] = square as u64;
+                position.king_square[piece.color().index()] = square;
             }
         }
